requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
google-cloud-tasks==2.16.5
google-cloud-storage
pydantic==2.9.2
